Move enums to GodotcordRelationship
